+++ ci/gpu/build.sh
@@ -74,6 +74,5 @@ cd $WORKSPACE/cpp/build
 GTEST_OUTPUT="xml:${WORKSPACE}/test-results/" gtests/GDFGRAPH_TEST
 
 logger "Python py.test for cuGraph..."
-cd $WORKSPACE
-tar -zxvf cpp/src/tests/datasets.tar.gz -C /
+cd $WORKSPACE/python
 py.test --cache-clear --junitxml=${WORKSPACE}/junit-cugraph.xml -v