+++ CHANGELOG.md
@@ -9,8 +9,10 @@
 ## Improvements
 - PR #157 Removed cudatoolkit dependency in setup.py
 - PR #185 Update docs version
+- PR #194 Open source nvgraph in cugraph repository #194
 - PR #190 Added a copy option in graph creation
 - PR #196 Fix typos in readme intro
+- PR #203 Added small datasets directly in the repo 
 
 
 ## Bug Fixes